requirements.txt
asgiref==3.5.0
attrs==23.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==4.0.2
django-limits==0.0.6
django-markdown-deux==1.0.5
django-ordered-model==3.5
django-ratelimit==3.0.1
django-stubs==5.0.0
django-stubs-ext==5.0.0
fancycompleter==0.9.1
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib-metadata==4.11.0
Markdown==3.3.6
markdown2==2.4.2
msgpack==1.0.8
pdbpp==0.10.3
Pillow==9.0.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyrepl==0.9.0
python-dotenv==1.0.1
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9
sqlparse==0.4.2
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18 ; python_version >= "3.9" and python_version < "4.0"
wmctrl==0.5
zipp==3.7.0
